--- Pocket-Sphinx-Parrot.py
@@ -1,2 +1,0 @@
-from pocketsphinx import LiveSpeech
-for phrase in LiveSpeech(): print(phrase)
