--- Responsibilities
@@ -1,27 +1,0 @@
-Plan 2020:
-
-Brandon
-
-Daniel M
-- Text input parser
-
-Dan D
-- Front end
-
-Ken
-- Processor (math commands)
-
-Matt
-
-Miguel
-- Processor (existing commands)
-
-Yehun
-- Definition for math commands (add to commands document)
-
-Unassigned tasks:
-- Detect the platform
-- Graphics engine (mobile vs desktop)
-- Conditional and flow control commands
-- Store commands vs. immediate execution of commands.
-- File read / write
