--- VariableTest.py
@@ -1,22 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Apr  4 08:50:52 2019
-
-@author: tc595
-"""
-
-import SIMPL_Processor
-
-pr = SIMPL_Processor.SIMPL_Processor()
-
-pr.create_variable("hello")
-pr.create_variable("world")
-
-
-pr.assign_value("hello","hello")
-pr.assign_value("world","world")
-
-print (pr.join("hello","world"))
-
-pr.run_project
-
