+++ pocker-sphinx/Pocket-Sphinx-Parrot.py
@@ -1,0 +1,3 @@
+from pocketsphinx import LiveSpeech
+for phrase in LiveSpeech():
+    print(phrase)
