+++ projects/TestOne.SIMPL
@@ -1,0 +1,17 @@
+OPEN NEW PROJECT Test One
+CREATE VARIABLE First
+CREATE VARIABLE Second
+CREATE VARIABLE Third
+SET First TO Hello
+SET Second TO World
+JOIN Second TO First AS Third
+COMMENT
+This should be the most basic type of program that can be created using
+only your voice.
+END COMMENT
+PRINT Third
+SAY VARIABLE Third
+SAVE PROJECT Test One
+CHANGE LINE Six TO SET Second TO Earth
+SAVE PROJECT Test One
+RUN PROJECT
